--- src/services/IService.ts
@@ -1,4 +1,0 @@
-export interface IService {
-  onResponse: (route: Buffer[], response: Buffer) => void;
-  request(route: Buffer[], operation: any): void;
-}
